+++ src/main/scala/krs/eligibility/api/ApiModule.scala
@@ -1,0 +1,7 @@
+package krs.eligibility.api
+
+import krs.eligibility.domain.{ DomainModule }
+
+trait ApiModule { this: DomainModule =>
+  val eligibilityApi = EligibilityApiImpl()
+}
