+++ src/main/scala/krs/eligibility/api/EligibilityApi.scala
@@ -1,0 +1,18 @@
+package krs.eligibility.api
+
+import com.twitter.util.{ Future }
+import krs.user.domain.{ User }
+import krs.partner.domain.{ Offer }
+
+import krs.eligibility.domain._
+
+trait EligibilityApi {
+  def filterEligible(user: User, offers: Seq[Offer]): Future[Seq[Offer]]
+}
+
+case class EligibilityApiImpl() extends EligibilityApi {
+  def filterEligible(user: User, offers: Seq[Offer]): Future[Seq[Offer]] = {
+    val filteredOffers = offers.filter(offer => OfferSystem.isEligible(user, offer))
+    Future.value(filteredOffers)
+  }
+}
