+++ src/main/scala/krs/eligibility/domain/DomainModule.scala
@@ -1,0 +1,7 @@
+package krs.eligibility.domain
+
+import krs.partner.api.{ PartnerApi }
+
+trait DomainModule {
+  def partnerRepository: PartnerApi
+}
