+++ src/main/scala/krs/eligibility/infrastructure/InfrastructureModule.scala
@@ -1,0 +1,10 @@
+package krs.eligibility.infrastructure
+
+import krs.eligibility.api._
+import krs.partner.infrastructure.{ Injector }
+
+trait InfrastructureModule { this: ApiModule =>
+
+  val partnerRepository = new Injector().partnerApi
+
+}
