+++ src/main/scala/krs/eligibility/infrastructure/Injector.scala
@@ -1,0 +1,6 @@
+package krs.eligibility.infrastructure
+
+import krs.eligibility.domain.{ DomainModule }
+import krs.eligibility.api.{ ApiModule }
+
+class Injector extends InfrastructureModule with ApiModule with DomainModule
