+++ src/main/scala/krs/partner/api/ApiModule.scala
@@ -3,5 +3,5 @@ package krs.partner.api
 import krs.partner.domain.{ DomainModule }
 
 trait ApiModule { this: DomainModule =>
-  val partnerApi = PartnerApi(partnerRepository)
+  val partnerApi = PartnerApiImpl(partnerRepository)
 }
