+++ src/main/scala/krs/partner/api/PartnerApi.scala
@@ -3,11 +3,11 @@ package krs.partner.api
 import com.twitter.util.{ Future }
 import krs.partner.domain._
 
-trait PartnerApiTrait {
+trait PartnerApi {
   def getOffers(creditScore: Int): Future[Seq[Offer]]
 }
 
-case class PartnerApi(repository: PartnerRepository) extends PartnerApiTrait {
+case class PartnerApiImpl(repository: PartnerRepository) extends PartnerApi {
   def getOffers(creditScore: Int): Future[Seq[Offer]] = {
     val offers = repository.loadOffers
     val filteredOffers = PartnerSystem().filterOffers(offers, creditScore)
