+++ src/main/scala/krs/partner/infrastructure/Injector.scala
@@ -1,0 +1,6 @@
+package krs.partner.infrastructure
+
+import krs.partner.domain.{ DomainModule }
+import krs.partner.api.{ ApiModule }
+
+class Injector extends InfrastructureModule with ApiModule with DomainModule
