+++ src/main/scala/krs/user/api/ApiModule.scala
@@ -1,8 +1,11 @@
 package krs.user.api
 
 import krs.user.domain.{ DomainModule }
+import krs.partner.api.{ PartnerApi }
+import krs.eligibility.api.{ EligibilityApi }
 
 trait ApiModule { this: DomainModule =>
-  val partnerApi: krs.partner.api.PartnerApiTrait
-  val userApi = UserApi(repository, partnerApi)
+  val partnerRepository: PartnerApi
+  val eligibilityApi: EligibilityApi
+  val userApi = UserApi(repository, partnerRepository, eligibilityApi)
 }
