+++ src/main/scala/krs/user/infrastructure/InfrastructureModule.scala
@@ -1,16 +1,11 @@
 package krs.user.infrastructure
 
 import krs.user.api._
-
-object PartnerModule
-  extends krs.partner.infrastructure.InfrastructureModule
-  with krs.partner.api.ApiModule
-  with krs.partner.domain.DomainModule
+import krs.partner.infrastructure.{ Injector }
 
 trait InfrastructureModule { this: ApiModule =>
 
   val repository = UserRepositoryMemory()
-  val partnerModule = PartnerModule
-  val partnerApi = partnerModule.partnerApi
-
+  val partnerRepository = new Injector().partnerApi
+  val eligibilityApi = new krs.eligibility.infrastructure.Injector().eligibilityApi
 }
